重构 RandomMeasAdd：expect_instance 拆分为 shadow/hamming 子目录、jackknife 目录化、更新测试与 qmeas 配置

+++ theory/physics/contemporary_physics/quant_comp_note/03_tools_practice/RandomMeasAdd/test/get_data.py
@@ -25,7 +25,7 @@ setting_runs = [
     SettingRun(num_settings=3**6, num_shots=1024),
 ]
 aer_opts = AerOptions(
-    method="matrix_product_state", device="CPU", precision="single", mitigation=True
+    method="matrix_product_state", device="CPU", precision="double", mitigation=True
 )
 
 for scheme_name, meas_indices in schemes.items():
